pyproject.toml
[project]
name = "codespeak"
version = "0.1.0"
description = "Modern CLI tool for managing code evaluation workflows"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "coverage",
    "click>=8.1.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.8.4",
]

[project.scripts]
codespeak = "codespeak.cli:cli"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
